+++ helpers/log_helper.py
@@ -13,16 +13,17 @@ levels = {
 
 def create_log(module):
     logger = logging.getLogger(module)
-    console_log = logging.StreamHandler(stream=sys.stdout)
+    if not logger.hasHandlers():
+        console_log = logging.StreamHandler(stream=sys.stdout)
 
-    log_level = os.environ.get('LOG_LEVEL', 'warning').lower()
+        log_level = os.environ.get('LOG_LEVEL', 'warning').lower()
 
-    logger.setLevel(levels[log_level])
-    console_log.setLevel(levels[log_level])
+        logger.setLevel(levels[log_level])
+        console_log.setLevel(levels[log_level])
 
-    formatter = logging.Formatter('%(asctime)s | %(name)s | %(levelname)s: %(message)s')  # noqa: E501
-    console_log.setFormatter(formatter)
+        formatter = logging.Formatter('%(asctime)s | %(name)s | %(levelname)s: %(message)s')  # noqa: E501
+        console_log.setFormatter(formatter)
 
-    logger.addHandler(console_log)
+        logger.addHandler(console_log)
 
     return logger