+++ helpers/query_helper.py
@@ -5,8 +5,8 @@ _NEW_PATRONS_QUERY = '''
         x.id, ptype_code, pcode3, home_library_code,
         TRIM(city), TRIM(region), TRIM(postal_code), TRIM(addr1),
         (activity_gmt AT TIME ZONE 'EST')::DATE,
-        (record_last_updated_gmt AT TIME ZONE 'EST')::DATE,
         deletion_date_gmt,
+        (record_last_updated_gmt AT TIME ZONE 'EST')::DATE,
         creation_date_gmt
     FROM (
         SELECT
@@ -28,16 +28,15 @@ _UPDATED_PATRONS_QUERY = '''
         x.id, ptype_code, pcode3, home_library_code,
         TRIM(city), TRIM(region), TRIM(postal_code), TRIM(addr1),
         (activity_gmt AT TIME ZONE 'EST')::DATE,
-        (creation_date_gmt AT TIME ZONE 'EST')::DATE,
         deletion_date_gmt,
+        (creation_date_gmt AT TIME ZONE 'EST')::DATE,
         record_last_updated_gmt
     FROM (
         SELECT
             id, record_last_updated_gmt, deletion_date_gmt, creation_date_gmt
         FROM sierra_view.record_metadata
         WHERE record_type_code = 'p'
-            AND record_last_updated_gmt >= '{cached_creation_dt}'
-            AND id NOT IN ({new_patron_ids})
+            AND record_last_updated_gmt >= '{cached_update_dt}'
         ORDER BY record_last_updated_gmt
         LIMIT {limit}) x
     LEFT JOIN sierra_view.patron_record_address
@@ -53,9 +52,21 @@ _DELETED_PATRONS_QUERY = '''
     FROM sierra_view.record_metadata
     WHERE record_type_code = 'p'
         AND deletion_date_gmt >= '{cached_deletion_date}'
-        AND id NOT IN ({processed_patron_ids})
     ORDER BY deletion_date_gmt
     LIMIT {limit};'''
+
+_REDSHIFT_ADDRESS_QUERY = '''
+    SELECT address_hash, patron_id, geoid
+    FROM public.{redshift_table}
+    WHERE address_hash IN ({address_hashes})
+'''
+
+_REDSHIFT_PATRON_QUERY = '''
+    SELECT patron_id, address_hash, postal_code, geoid, creation_date_et,
+        circ_active_date_et, ptype_code, pcode3, patron_home_library_code
+    FROM public.{redshift_table}
+    WHERE patron_id IN ({patron_ids})
+'''
 
 
 def build_new_patrons_query(creation_dt_start):
@@ -65,16 +76,26 @@ def build_new_patrons_query(creation_dt_start):
         total_limit=int(os.environ['SIERRA_BATCH_SIZE']) * 2)
 
 
-def build_updated_patrons_query(update_dt_start, new_patron_ids):
+def build_updated_patrons_query(update_dt_start):
     return _UPDATED_PATRONS_QUERY.format(
         cached_update_dt=update_dt_start,
-        new_patron_ids=str(new_patron_ids)[1:-1],
         limit=os.environ['SIERRA_BATCH_SIZE'],
         total_limit=int(os.environ['SIERRA_BATCH_SIZE']) * 2)
 
 
-def build_deleted_patrons_query(deletion_date_start, processed_patron_ids):
+def build_deleted_patrons_query(deletion_date_start):
     return _DELETED_PATRONS_QUERY.format(
         cached_deletion_date=deletion_date_start,
-        processed_patron_ids=str(processed_patron_ids)[1:-1],
         limit=os.environ['SIERRA_BATCH_SIZE'])
+
+
+def build_redshift_address_query(address_hashes):
+    return _REDSHIFT_ADDRESS_QUERY.format(
+        redshift_table=os.environ['REDSHIFT_TABLE'],
+        address_hashes=address_hashes)
+
+
+def build_redshift_patron_query(patron_ids):
+    return _REDSHIFT_PATRON_QUERY.format(
+        redshift_table=os.environ['REDSHIFT_TABLE'],
+        patron_ids=patron_ids)